Add terraform version controls
Add globals to set minimum and latest Terraform versions.

TERRAFORM_VERSION_MINIMUM: The minimum required version. If there is an
existing binary that matches the minor version and is equal to or
greater than this value then a new binary will not be downloaded.

TERRAFORM_VERSION_LATEST:  The version that will be downloaded if there
is no existing binary, or if the existing binary does not meet the
minimum version requirement.