Fix Terraform formatting of 'map' type arguments
Terraform arguments need to be defined using an '='.

Currently the Terraform provider 'map' type arguments (translated to
'dict' in Python) are formatted omitting the '=' resulting in Terraform
interpreting it as a block, and raising an 'Error: Unsupported block
type' exception.

There's also an indentation mistake for the 'dict' type arguments which
didn't cause any errors but looked messy.