Fix network ID
The return data structure for 'pi_network_existing_output.resource' and
'pi_network_create_output.resource' are not the same. In particular,
the "existing" output returns an 'id' with just the network id, while
the "create" output returns an 'id' with
'{pc_cloud_instance_id}/{network_id}' along with a separate
'network_id'. Only the network id is needed, so the fact is updated to
container just that.