Simple VM SSH (VPC) example update (#99)
* Update 'simple-vm-ssh' example destroy playbook

Previously, the destroy.yml playbook assumed that fact caching was
enabled, and that resource information was available from when the
create.yml playbook was run.

The playbook has been updated to get resource information from the
PowerVS service itself (using resource names). This approach:
- does not require fact caching
- does not require running from the same host that ran create.yml
- is much easier to modify and extend

* Update example image

Ubuntu 16.04 is EOL

* Fix formatting

* Set resource names in vars.yml

* Lookup resources instead of relying on fact cache

* Update README vars description

* Add example creating multiple VPC VSIs

* Remove path to 'ansible_ssh_private_key_file'

This was used for testing and was inadvertantly left in place.

* User python split method instead of Jinja2 filter

Older versions of Ansible do not support the 'split' Jinja2 filter.
Using the Python method directly is a better approach.